+++ src/ReturnTypes/ModelExtension.php
@@ -9,27 +9,21 @@ use Illuminate\Database\Eloquent\Collection;
 use Illuminate\Database\Eloquent\Model;
 use Illuminate\Database\Query\Builder as QueryBuilder;
 use NunoMaduro\Larastan\Methods\BuilderHelper;
-use NunoMaduro\Larastan\Methods\ModelTypeHelper;
-use NunoMaduro\Larastan\Methods\Pipes\Mixins;
 use PhpParser\Node\Expr\StaticCall;
 use PHPStan\Analyser\Scope;
-use PHPStan\Reflection\FunctionVariantWithPhpDocs;
 use PHPStan\Reflection\MethodReflection;
+use PHPStan\Reflection\ParametersAcceptorSelector;
 use PHPStan\Reflection\ReflectionProvider;
 use PHPStan\Type\DynamicStaticMethodReturnTypeExtension;
 use PHPStan\Type\Generic\GenericObjectType;
 use PHPStan\Type\ObjectType;
 use PHPStan\Type\Type;
-use ReflectionClass;
 
 /**
  * @internal
  */
 final class ModelExtension implements DynamicStaticMethodReturnTypeExtension
 {
-    /** @var Mixins */
-    private $mixins;
-
     /** @var BuilderHelper */
     private $builderHelper;
 
@@ -38,11 +32,9 @@ final class ModelExtension implements DynamicStaticMethodReturnTypeExtension
 
     /**
      * @param BuilderHelper $builderHelper
-     * @param Mixins|null   $mixins
      */
-    public function __construct(ReflectionProvider $reflectionProvider, BuilderHelper $builderHelper, Mixins $mixins = null)
+    public function __construct(ReflectionProvider $reflectionProvider, BuilderHelper $builderHelper)
     {
-        $this->mixins = $mixins ?? new Mixins();
         $this->builderHelper = $builderHelper;
         $this->reflectionProvider = $reflectionProvider;
     }
@@ -83,32 +75,7 @@ final class ModelExtension implements DynamicStaticMethodReturnTypeExtension
         $method = $methodReflection->getDeclaringClass()
             ->getMethod($methodReflection->getName(), $scope);
 
-        $variants = $method->getVariants();
-        $returnType = $variants[0]->getReturnType();
-
-        /*
-         * If the method returns a static type, we instruct phpstan that
-         * "static" points to the concrete class model.
-         */
-        if ($methodCall->class instanceof \PhpParser\Node\Name && $variants[0] instanceof FunctionVariantWithPhpDocs) {
-            $className = $methodCall->class->toString();
-            if (class_exists($className)) {
-                $classReflection = new ReflectionClass($className);
-                $isValidInstance = false;
-                foreach ($this->mixins->getMixinsFromClass(
-                    $this->reflectionProvider,
-                    $this->reflectionProvider->getClass(Collection::class)
-                ) as $mixin) {
-                    if ($isValidInstance = $classReflection->isSubclassOf($mixin)) {
-                        break;
-                    }
-                }
-
-                if ($isValidInstance) {
-                    $returnType = ModelTypeHelper::replaceStaticTypeWithModel($returnType, $className);
-                }
-            }
-        }
+        $returnType = ParametersAcceptorSelector::selectSingle($method->getVariants())->getReturnType();
 
         if ((count(array_intersect([EloquentBuilder::class, QueryBuilder::class], $returnType->getReferencedClasses())) > 0)
             && $methodCall->class instanceof \PhpParser\Node\Name