+++ tests/Features/Methods/ModelExtension.php
@@ -10,7 +10,6 @@ use Illuminate\Database\Eloquent\Builder;
 use Illuminate\Database\Eloquent\Collection;
 use Illuminate\Database\Eloquent\Model;
 use Illuminate\Database\Eloquent\Relations\HasMany;
-use Illuminate\Database\Eloquent\Relations\Relation;
 use Illuminate\Foundation\Http\FormRequest;
 
 class ModelExtension
@@ -214,7 +213,7 @@ class ModelExtension
 
     public function testWithAcceptsArrayOfClosures(): ?User
     {
-        return User::with(['accounts' => function (Relation $relation) {
+        return User::with(['accounts' => function (HasMany $relation) {
             return $relation->where('active', true);
         }])->find(1);
     }